+++ src/modify/delivel.level-grind/PetLayer.cpp
@@ -1,27 +1,33 @@
 #include <Geode/Geode.hpp>
 #include <alphalaneous.alphas_geode_utils/include/ObjectModify.hpp>
+#include <cue/RepeatingBackground.hpp>
 #include "../../Utils.hpp"
 
 using namespace geode::prelude;
 
-class $nodeModify(MyBetterAchievementLayer, BetterAchievementLayer){
+class $nodeModify(MyPetLayer, PetLayer){
 
     void modify(){
-        
+
         auto DayAndNightSystem = DayAndNightSystem::create();
         DayAndNightSystem->setID("events"_spr);
-        this->addChild(DayAndNightSystem, -2);
+        this->addChild(DayAndNightSystem, -1);
         
         auto DayAndNightSystemOverlay = DayAndNightSystemOverlay::create();
         DayAndNightSystemOverlay->setID("screen-overlay"_spr);
         this->addChild(DayAndNightSystemOverlay, 106);
 
         if (DayAndNightSystem::events > 0.99f){
-            if (auto bg = this->getChildByID("content-background")){
+
+            if (auto rBG = getChildByType<cue::RepeatingBackground>(0)){   
+                rBG->setVisible(false);
+            }
+            else if (auto bg = this->getChildByID("bg")){
 			    bg->setVisible(false);
 		    }
-        }
 
+        }
+        
     }
 
 };